ape-gpt-cli/requirements.txt
openai>=0.10.2
PyYAML>=5.4.1